requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.5
django-cors-headers==3.4.0
firebase-admin==4.3.0
google-auth==1.35.0
fcm-django==0.3.4
pyfcm==1.4.7
gunicorn==20.0.4